pyproject.toml
[project]
name = "postgis-site-selector"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "requests>=2.33.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]
